// src/generated/source-data.ts
export const constants = {
  "KNPLAY": 10,
  "KNBASE": 10,
  "KGALV": 75,
  "KSID": 25,
  "KGALH": 75,
  "BRDSIZ": 1875,
  "KNPLNT": 60,
  "KNDEV": 9,
  "KPASS": "*MINK",
  "KNMSG": 32,
  "MSGLEN": 17,
  "KNGRP": 7,
  "KNHIT": 64,
  "KNCMD": 33,
  "KNPCMD": 16,
  "KNPOIN": 8,
  "KNXTR": 8,
  "KSCMD": 2,
  "KMAXTK": 15,
  "KNTTY": 8,
  "KRANGE": 10,
  "KCRIT": 3000,
  "KENDAM": 25000,
  "KWAIT": 0,
  "KCMDTM": 2000,
  "KQLEN": 10,
  "MAXINT": 34359738367,
  "LOCSIZ": 200,
  "KEOL": -1,
  "KNUL": 0,
  "KINT": 1,
  "KFLT": 2,
  "KALF": 3,
  "KVPOS": 1,
  "KHPOS": 2,
  "KNTURN": 3,
  "KSPCON": 4,
  "KNTORP": 5,
  "KSHCON": 6,
  "KLFSUP": 7,
  "KSNRGY": 8,
  "KSDAM": 9,
  "KSSHPC": 10,
  "KPEDAM": 1,
  "KPEKIL": 2,
  "KPBDAM": 3,
  "KPPCAP": 4,
  "KPBBAS": 5,
  "KPRKIL": 6,
  "KNSDES": 7,
  "KNPDES": 8,
  "KDSHLD": 1,
  "KDWARP": 2,
  "KDIMP": 3,
  "KDLIFE": 4,
  "KDTORP": 5,
  "KDPHAS": 6,
  "KDCOMP": 7,
  "KDRAD": 8,
  "KNJBST": 9,
  "KJOB": 1,
  "KNAM1": 2,
  "KNAM2": 3,
  "KPPN": 4,
  "KTTYN": 5,
  "KTTYSP": 6,
  "KTTYTP": 7,
  "KJOBTM": 8,
  "KRUNTM": 9,
  "NAMBIT": 1,
  "PRTBIT": 2,
  "SHPBIT": 4,
  "BASBIT": 8,
  "PLNBIT": 16,
  "CRDBIT": 32,
  "FEDBIT": 64,
  "EMPBIT": 128,
  "ROMBIT": 256,
  "NEUBIT": 512,
  "CAPBIT": 1024,
  "ALLBIT": 2048,
  "RNGBIT": 4096,
  "CLSBIT": 8192,
  "SUMBIT": 16384,
  "LSTBIT": 32768,
  "OUTMSK": 49152,
  "IRNBIT": 1,
  "ISRBIT": 2,
  "IGMBIT": 4,
  "ORNBIT": 8,
  "PASBIT": 16,
  "KNOBIT": 32,
  "SIDMSK": 1984,
  "OBJMSK": 63,
  "GREEN": 1,
  "YELLOW": 2,
  "RED": 3,
  "SHORT": -1,
  "MEDIUM": 0,
  "LONG": 1,
  "KREL": -1,
  "KBOTH": 0,
  "KABS": 1,
  "LSTCMD": 0,
  "SUMCMD": 1,
  "BASCMD": 2,
  "PLNCMD": 3,
  "TARCMD": 4,
  "DXMPTY": 0,
  "DXFSHP": 1,
  "DXESHP": 2,
  "DXFBAS": 3,
  "DXEBAS": 4,
  "DXROM": 5,
  "DXNPLN": 6,
  "DXFPLN": 7,
  "DXEPLN": 8,
  "DXSTAR": 9,
  "DXBHOL": 10
} as const;
// ...
export const honorRollText = [
  {
    "text": "\r\n\r\n\r\n--------------\r\n\r\nThe DECWAR Honor Roll\r\n\r\n(* indicates Missing in Action)\r\n\r\n",
    "file": "WARMAC.MAC",
    "line": 5902
  },
  {
    "text": "(**** non-paying users ****)\r\n\r\n",
    "file": "WARMAC.MAC",
    "line": 5916
  },
  {
    "text": "\r\nThe Federation has awarded the\r\nfollowing Captains the Emerald\r\nStar Cluster for outstanding\r\nservice:\r\n\r\n",
    "file": "WARMAC.MAC",
    "line": 5947
  },
  {
    "text": "\r\n\r\nThe Golden Galaxy Medal has been\r\nawarded in memory of these brave\r\nCaptains:\r\n\r\n",
    "file": "WARMAC.MAC",
    "line": 5961
  },
  {
    "text": "\r\n\r\nThe following Captains have served\r\ntheir Empire well:\r\n\r\n",
    "file": "WARMAC.MAC",
    "line": 5975
  },
  {
    "text": "\r\n\r\nThe Distinguished Service Cross\r\nhas been posthumously awarded\r\nto the following Captains for\r\ntheir outstanding service:\r\n\r\n",
    "file": "WARMAC.MAC",
    "line": 5988
  },
  {
    "text": "\r\nCaptain        Service # Credits",
    "file": "WARMAC.MAC",
    "line": 6007
  },
  {
    "text": " Ship        Runtm Date",
    "file": "WARMAC.MAC",
    "line": 6014
  },
  {
    "text": "   ",
    "file": "WARMAC.MAC",
    "line": 6077
  },
  {
    "text": "/",
    "file": "WARMAC.MAC",
    "line": 6083
  },
  {
    "text": "/",
    "file": "WARMAC.MAC",
    "line": 6087
  }
] as const;

export const clearStatisticsText = [
  {
    "text": "\r\nZapping statistics logs....",
    "file": "WARMAC.MAC",
    "line": 6186
  },
  {
    "text": "\r\nFinished!\r\n",
    "file": "WARMAC.MAC",
    "line": 6210
  },
  {
    "text": "\r\n\r\nCan't open file for output!\r\n\r\n",
    "file": "WARMAC.MAC",
    "line": 6216
  }
] as const;
// ...
export const gripeText = [
  {
    "text": "\r\nYou are not permitted to GRIPE\r\nwhile under RED alert!\r\n",
    "file": "WARMAC.MAC",
    "line": 4719
  },
  {
    "text": "Enter gripe, end with ^Z\r\n",
    "file": "WARMAC.MAC",
    "line": 4733
  },
  {
    "text": "[Only 2 more message lines allowed]\r\n",
    "file": "WARMAC.MAC",
    "line": 4759
  },
  {
    "text": "[Too many lines -- end of gripe]\r\n",
    "file": "WARMAC.MAC",
    "line": 4761
  },
  {
    "text": "**** Data out of bounds ****\r\n**** Command line:\r\n",
    "file": "WARMAC.MAC",
    "line": 4774
  },
  {
    "text": "*** PDL: ",
    "file": "WARMAC.MAC",
    "line": 4845
  },
  {
    "text": "*** HITQL-1:",
    "file": "WARMAC.MAC",
    "line": 4865
  },
  {
    "text": "*** LOKTAB:",
    "file": "WARMAC.MAC",
    "line": 4881
  },
  {
    "text": "----------",
    "file": "WARMAC.MAC",
    "line": 4920
  },
  {
    "text": "%Can't write DECWAR.GRP\r\n",
    "file": "WARMAC.MAC",
    "line": 4932
  },
  {
    "text": "%DECWAR.GRP being modified; trying again\r\n",
    "file": "WARMAC.MAC",
    "line": 4933
  },
  {
    "text": "%Can't get core to read DECWAR.GRP\r\n",
    "file": "WARMAC.MAC",
    "line": 4951
  },
  {
    "text": "%Can't read DECWAR.GRP\r\n",
    "file": "WARMAC.MAC",
    "line": 4958
  },
  {
    "text": "%Can't write DECWAR.GRP\r\n",
    "file": "WARMAC.MAC",
    "line": 4969
  },
  {
    "text": "%Can't get more core\r\n",
    "file": "WARMAC.MAC",
    "line": 4992
  }
] as const;